Separe la logica de validar transaccion de la de agregar transaccion. Agregue numero a los bancos, cambie atributos de dto.

+++ Services/CuentaService.cs
@@ -44,7 +44,7 @@ namespace Broker.Services
             return cuenta.cuentaId;
         }
 
-        public async Task<CuentaDto> buscarCuenta(double numero)
+        public async Task<CuentaDto> buscarCuenta(long numero)
         {
             // Realiza la búsqueda de transaccion por numero
             var cuenta = await _context.Cuenta
